+++ contracts/liquidity_hub/pool-network/stableswap_3pool/src/commands.rs
@@ -16,6 +16,7 @@ use white_whale::pool_network::asset::{
 use white_whale::pool_network::denom::{Coin, MsgBurn, MsgMint};
 #[cfg(feature = "osmosis_token_factory")]
 use white_whale::pool_network::denom_osmosis::{Coin, MsgBurn, MsgMint};
+use white_whale::pool_network::swap;
 use white_whale::pool_network::trio::{Config, Cw20HookMsg, FeatureToggle, PoolFee, RampAmp};
 
 use crate::error::ContractError;
@@ -387,24 +388,16 @@ pub fn swap(
     let ask_pool: Asset;
     let offer_pool: Asset;
     let unswapped_pool: Asset;
-    let ask_decimal: u8;
-    let offer_decimal: u8;
 
     if ask_asset.equal(&pools[0].info) {
         if offer_asset.info.equal(&pools[1].info) {
             ask_pool = pools[0].clone();
             offer_pool = pools[1].clone();
             unswapped_pool = pools[2].clone();
-
-            ask_decimal = trio_info.asset_decimals[0];
-            offer_decimal = trio_info.asset_decimals[1];
         } else if offer_asset.info.equal(&pools[2].info) {
             ask_pool = pools[0].clone();
             offer_pool = pools[2].clone();
             unswapped_pool = pools[1].clone();
-
-            ask_decimal = trio_info.asset_decimals[0];
-            offer_decimal = trio_info.asset_decimals[2];
         } else {
             return Err(ContractError::AssetMismatch {});
         }
@@ -413,16 +406,10 @@ pub fn swap(
             ask_pool = pools[1].clone();
             offer_pool = pools[0].clone();
             unswapped_pool = pools[2].clone();
-
-            ask_decimal = trio_info.asset_decimals[1];
-            offer_decimal = trio_info.asset_decimals[0];
         } else if offer_asset.info.equal(&pools[2].info) {
             ask_pool = pools[1].clone();
             offer_pool = pools[2].clone();
             unswapped_pool = pools[0].clone();
-
-            ask_decimal = trio_info.asset_decimals[1];
-            offer_decimal = trio_info.asset_decimals[2];
         } else {
             return Err(ContractError::AssetMismatch {});
         }
@@ -431,16 +418,10 @@ pub fn swap(
             ask_pool = pools[2].clone();
             offer_pool = pools[0].clone();
             unswapped_pool = pools[1].clone();
-
-            ask_decimal = trio_info.asset_decimals[2];
-            offer_decimal = trio_info.asset_decimals[0];
         } else if offer_asset.info.equal(&pools[1].info) {
             ask_pool = pools[2].clone();
             offer_pool = pools[1].clone();
             unswapped_pool = pools[0].clone();
-
-            ask_decimal = trio_info.asset_decimals[2];
-            offer_decimal = trio_info.asset_decimals[1];
         } else {
             return Err(ContractError::AssetMismatch {});
         }
@@ -478,7 +459,7 @@ pub fn swap(
         .checked_add(swap_computation.burn_fee_amount)?;
 
     // check max spread limit if exist
-    helpers::assert_max_spread(
+    swap::assert_max_spread(
         belief_price,
         max_spread,
         offer_asset.amount,
