+++ contracts/liquidity_hub/pool-network/stableswap_3pool/src/error.rs
@@ -19,9 +19,6 @@ pub enum ContractError {
     #[error("Invalid zero amount")]
     InvalidZeroAmount {},
 
-    #[error("Spread limit exceeded")]
-    MaxSpreadAssertion {},
-
     #[error("Slippage tolerance exceeded")]
     MaxSlippageAssertion {},
