+++ contracts/liquidity_hub/pool-network/terraswap_pair/src/error.rs
@@ -34,9 +34,6 @@ pub enum ContractError {
     #[error("Invalid zero amount")]
     InvalidZeroAmount {},
 
-    #[error("Spread limit exceeded")]
-    MaxSpreadAssertion {},
-
     #[error("Slippage tolerance exceeded")]
     MaxSlippageAssertion {},
