+++ contracts/liquidity_hub/pool-network/terraswap_pair/src/helpers.rs
@@ -1,12 +1,11 @@
 use std::ops::Mul;
-use std::str::FromStr;
 
 use cosmwasm_schema::cw_serde;
 #[cfg(any(feature = "token_factory", feature = "osmosis_token_factory"))]
 use cosmwasm_std::CosmosMsg;
 use cosmwasm_std::{
-    to_binary, Decimal, Decimal256, DepsMut, Env, Fraction, ReplyOn, Response, StdError, StdResult,
-    Storage, SubMsg, Uint128, Uint256, WasmMsg,
+    to_binary, Decimal, Decimal256, DepsMut, Env, ReplyOn, Response, StdError, StdResult, Storage,
+    SubMsg, Uint128, Uint256, WasmMsg,
 };
 use cw20::MinterResponse;
 use cw_storage_plus::Item;
@@ -320,58 +319,6 @@ pub struct OfferAmountComputation {
     pub swap_fee_amount: Uint128,
     pub protocol_fee_amount: Uint128,
     pub burn_fee_amount: Uint128,
-}
-
-/// Default swap slippage in case max_spread is not specified
-pub const DEFAULT_SLIPPAGE: &str = "0.01";
-/// Cap on the maximum swap slippage that is allowed. If max_spread goes over this limit, it will
-/// be capped to this value.
-pub const MAX_ALLOWED_SLIPPAGE: &str = "0.5";
-
-/// If `belief_price` and `max_spread` both are given,
-/// we compute new spread else we just use pool network
-/// spread to check `max_spread`
-pub fn assert_max_spread(
-    belief_price: Option<Decimal>,
-    max_spread: Option<Decimal>,
-    offer_amount: Uint128,
-    return_amount: Uint128,
-    spread_amount: Uint128,
-) -> Result<(), ContractError> {
-    println!("assert_max_spread: belief_price: {:?}, max_spread: {:?}, offer_amount: {:?}, return_amount: {:?}, spread_amount: {:?}", belief_price, max_spread, offer_amount, return_amount, spread_amount);
-
-    let max_spread: Decimal256 = max_spread
-        .unwrap_or(Decimal::from_str(DEFAULT_SLIPPAGE)?)
-        .min(Decimal::from_str(MAX_ALLOWED_SLIPPAGE)?)
-        .into();
-
-    println!("max_spread: {:?}", max_spread);
-    println!(
-        "Decimal256::from_ratio(spread_amount, return_amount + spread_amount) {:?}",
-        Decimal256::from_ratio(spread_amount, return_amount + spread_amount)
-    );
-    println!(
-        "Decimal256::from_ratio(spread_amount, return_amount + spread_amount) > max_spread: {:?}",
-        Decimal256::from_ratio(spread_amount, return_amount + spread_amount) > max_spread
-    );
-
-    if let Some(belief_price) = belief_price {
-        let expected_return = offer_amount
-            * belief_price
-                .inv()
-                .ok_or_else(|| StdError::generic_err("Belief price can't be zero"))?;
-        let spread_amount = expected_return.saturating_sub(return_amount);
-
-        if return_amount < expected_return
-            && Decimal256::from_ratio(spread_amount, expected_return) > max_spread
-        {
-            return Err(ContractError::MaxSpreadAssertion {});
-        }
-    } else if Decimal256::from_ratio(spread_amount, return_amount + spread_amount) > max_spread {
-        return Err(ContractError::MaxSpreadAssertion {});
-    }
-
-    Ok(())
 }
 
 pub fn assert_slippage_tolerance(
