+++ contracts/liquidity_hub/pool-network/terraswap_router/schema/raw/execute.json
@@ -27,6 +27,16 @@
             "operations"
           ],
           "properties": {
+            "max_spread": {
+              "anyOf": [
+                {
+                  "$ref": "#/definitions/Decimal"
+                },
+                {
+                  "type": "null"
+                }
+              ]
+            },
             "minimum_receive": {
               "anyOf": [
                 {
@@ -68,6 +78,16 @@
             "operation"
           ],
           "properties": {
+            "max_spread": {
+              "anyOf": [
+                {
+                  "$ref": "#/definitions/Decimal"
+                },
+                {
+                  "type": "null"
+                }
+              ]
+            },
             "operation": {
               "$ref": "#/definitions/SwapOperation"
             },
@@ -215,6 +235,10 @@
         }
       },
       "additionalProperties": false
+    },
+    "Decimal": {
+      "description": "A fixed-point decimal value with 18 fractional digits, i.e. Decimal(1_000_000_000_000_000_000) == 1.0\n\nThe greatest possible value that can be represented is 340282366920938463463.374607431768211455 (which is (2^128 - 1) / 10^18)",
+      "type": "string"
     },
     "SwapOperation": {
       "oneOf": [
