--- contracts/liquidity_hub/vault-manager/schema/raw/query.json
@@ -1,205 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "title": "QueryMsg",
-  "description": "The query messages",
-  "oneOf": [
-    {
-      "description": "Retrieves the configuration of the manager.",
-      "type": "object",
-      "required": [
-        "config"
-      ],
-      "properties": {
-        "config": {
-          "type": "object",
-          "additionalProperties": false
-        }
-      },
-      "additionalProperties": false
-    },
-    {
-      "description": "Retrieves a vault given the asset_info.",
-      "type": "object",
-      "required": [
-        "vault"
-      ],
-      "properties": {
-        "vault": {
-          "type": "object",
-          "required": [
-            "asset_info"
-          ],
-          "properties": {
-            "asset_info": {
-              "$ref": "#/definitions/AssetInfo"
-            }
-          },
-          "additionalProperties": false
-        }
-      },
-      "additionalProperties": false
-    },
-    {
-      "description": "Retrieves the addresses for all the vaults.",
-      "type": "object",
-      "required": [
-        "vaults"
-      ],
-      "properties": {
-        "vaults": {
-          "type": "object",
-          "properties": {
-            "limit": {
-              "type": [
-                "integer",
-                "null"
-              ],
-              "format": "uint32",
-              "minimum": 0.0
-            },
-            "start_after": {
-              "type": [
-                "array",
-                "null"
-              ],
-              "items": {
-                "type": "integer",
-                "format": "uint8",
-                "minimum": 0.0
-              }
-            }
-          },
-          "additionalProperties": false
-        }
-      },
-      "additionalProperties": false
-    },
-    {
-      "description": "Retrieves the share of the assets stored in the vault that a given `lp_share` is entitled to.",
-      "type": "object",
-      "required": [
-        "share"
-      ],
-      "properties": {
-        "share": {
-          "type": "object",
-          "required": [
-            "lp_share"
-          ],
-          "properties": {
-            "lp_share": {
-              "$ref": "#/definitions/Asset"
-            }
-          },
-          "additionalProperties": false
-        }
-      },
-      "additionalProperties": false
-    },
-    {
-      "description": "Retrieves the [`Uint128`] amount that must be sent back to the contract to pay off a loan taken out.",
-      "type": "object",
-      "required": [
-        "payback_amount"
-      ],
-      "properties": {
-        "payback_amount": {
-          "type": "object",
-          "required": [
-            "asset"
-          ],
-          "properties": {
-            "asset": {
-              "$ref": "#/definitions/Asset"
-            }
-          },
-          "additionalProperties": false
-        }
-      },
-      "additionalProperties": false
-    },
-    {
-      "description": "Query the contract's ownership information",
-      "type": "object",
-      "required": [
-        "ownership"
-      ],
-      "properties": {
-        "ownership": {
-          "type": "object",
-          "additionalProperties": false
-        }
-      },
-      "additionalProperties": false
-    }
-  ],
-  "definitions": {
-    "Asset": {
-      "type": "object",
-      "required": [
-        "amount",
-        "info"
-      ],
-      "properties": {
-        "amount": {
-          "$ref": "#/definitions/Uint128"
-        },
-        "info": {
-          "$ref": "#/definitions/AssetInfo"
-        }
-      },
-      "additionalProperties": false
-    },
-    "AssetInfo": {
-      "description": "AssetInfo contract_addr is usually passed from the cw20 hook so we can trust the contract_addr is properly validated.",
-      "oneOf": [
-        {
-          "type": "object",
-          "required": [
-            "token"
-          ],
-          "properties": {
-            "token": {
-              "type": "object",
-              "required": [
-                "contract_addr"
-              ],
-              "properties": {
-                "contract_addr": {
-                  "type": "string"
-                }
-              },
-              "additionalProperties": false
-            }
-          },
-          "additionalProperties": false
-        },
-        {
-          "type": "object",
-          "required": [
-            "native_token"
-          ],
-          "properties": {
-            "native_token": {
-              "type": "object",
-              "required": [
-                "denom"
-              ],
-              "properties": {
-                "denom": {
-                  "type": "string"
-                }
-              },
-              "additionalProperties": false
-            }
-          },
-          "additionalProperties": false
-        }
-      ]
-    },
-    "Uint128": {
-      "description": "A thin wrapper around u128 that is using strings for JSON encoding/decoding, such that the full u128 range can be used for clients that convert JSON numbers to floats, like JavaScript and jq.\n\n# Examples\n\nUse `from` to create instances of this and `u128` to get the value out:\n\n``` # use cosmwasm_std::Uint128; let a = Uint128::from(123u128); assert_eq!(a.u128(), 123);\n\nlet b = Uint128::from(42u64); assert_eq!(b.u128(), 42);\n\nlet c = Uint128::from(70u32); assert_eq!(c.u128(), 70); ```",
-      "type": "string"
-    }
-  }
-}
