--- contracts/liquidity_hub/vault-manager/schema/raw/response_to_config.json
@@ -1,151 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "title": "Config",
-  "description": "Configuration for the contract (manager)",
-  "type": "object",
-  "required": [
-    "deposit_enabled",
-    "flash_loan_enabled",
-    "lp_token_type",
-    "vault_creation_fee",
-    "whale_lair_addr",
-    "withdraw_enabled"
-  ],
-  "properties": {
-    "deposit_enabled": {
-      "description": "If deposits are enabled",
-      "type": "boolean"
-    },
-    "flash_loan_enabled": {
-      "description": "If flash-loans are enabled",
-      "type": "boolean"
-    },
-    "lp_token_type": {
-      "description": "The type of LP token to use, whether a cw20 token or a token factory token",
-      "allOf": [
-        {
-          "$ref": "#/definitions/LpTokenType"
-        }
-      ]
-    },
-    "vault_creation_fee": {
-      "description": "The fee to create a new vault",
-      "allOf": [
-        {
-          "$ref": "#/definitions/Asset"
-        }
-      ]
-    },
-    "whale_lair_addr": {
-      "description": "The whale lair contract address",
-      "allOf": [
-        {
-          "$ref": "#/definitions/Addr"
-        }
-      ]
-    },
-    "withdraw_enabled": {
-      "description": "If withdrawals are enabled",
-      "type": "boolean"
-    }
-  },
-  "additionalProperties": false,
-  "definitions": {
-    "Addr": {
-      "description": "A human readable address.\n\nIn Cosmos, this is typically bech32 encoded. But for multi-chain smart contracts no assumptions should be made other than being UTF-8 encoded and of reasonable length.\n\nThis type represents a validated address. It can be created in the following ways 1. Use `Addr::unchecked(input)` 2. Use `let checked: Addr = deps.api.addr_validate(input)?` 3. Use `let checked: Addr = deps.api.addr_humanize(canonical_addr)?` 4. Deserialize from JSON. This must only be done from JSON that was validated before such as a contract's state. `Addr` must not be used in messages sent by the user because this would result in unvalidated instances.\n\nThis type is immutable. If you really need to mutate it (Really? Are you sure?), create a mutable copy using `let mut mutable = Addr::to_string()` and operate on that `String` instance.",
-      "type": "string"
-    },
-    "Asset": {
-      "type": "object",
-      "required": [
-        "amount",
-        "info"
-      ],
-      "properties": {
-        "amount": {
-          "$ref": "#/definitions/Uint128"
-        },
-        "info": {
-          "$ref": "#/definitions/AssetInfo"
-        }
-      },
-      "additionalProperties": false
-    },
-    "AssetInfo": {
-      "description": "AssetInfo contract_addr is usually passed from the cw20 hook so we can trust the contract_addr is properly validated.",
-      "oneOf": [
-        {
-          "type": "object",
-          "required": [
-            "token"
-          ],
-          "properties": {
-            "token": {
-              "type": "object",
-              "required": [
-                "contract_addr"
-              ],
-              "properties": {
-                "contract_addr": {
-                  "type": "string"
-                }
-              },
-              "additionalProperties": false
-            }
-          },
-          "additionalProperties": false
-        },
-        {
-          "type": "object",
-          "required": [
-            "native_token"
-          ],
-          "properties": {
-            "native_token": {
-              "type": "object",
-              "required": [
-                "denom"
-              ],
-              "properties": {
-                "denom": {
-                  "type": "string"
-                }
-              },
-              "additionalProperties": false
-            }
-          },
-          "additionalProperties": false
-        }
-      ]
-    },
-    "LpTokenType": {
-      "description": "The type of LP token to use, whether a cw20 token or a token factory token",
-      "oneOf": [
-        {
-          "type": "string",
-          "enum": [
-            "token_factory"
-          ]
-        },
-        {
-          "type": "object",
-          "required": [
-            "cw20"
-          ],
-          "properties": {
-            "cw20": {
-              "type": "integer",
-              "format": "uint64",
-              "minimum": 0.0
-            }
-          },
-          "additionalProperties": false
-        }
-      ]
-    },
-    "Uint128": {
-      "description": "A thin wrapper around u128 that is using strings for JSON encoding/decoding, such that the full u128 range can be used for clients that convert JSON numbers to floats, like JavaScript and jq.\n\n# Examples\n\nUse `from` to create instances of this and `u128` to get the value out:\n\n``` # use cosmwasm_std::Uint128; let a = Uint128::from(123u128); assert_eq!(a.u128(), 123);\n\nlet b = Uint128::from(42u64); assert_eq!(b.u128(), 42);\n\nlet c = Uint128::from(70u32); assert_eq!(c.u128(), 70); ```",
-      "type": "string"
-    }
-  }
-}
