--- contracts/liquidity_hub/vault-manager/schema/raw/response_to_payback_amount.json
@@ -1,100 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "title": "PaybackAssetResponse",
-  "description": "Response for the PaybackAmount query. Contains the amount that must be paid back to the contract if taken a flashloan.",
-  "type": "object",
-  "required": [
-    "asset_info",
-    "flash_loan_fee",
-    "payback_amount",
-    "protocol_fee"
-  ],
-  "properties": {
-    "asset_info": {
-      "description": "The asset info of the asset that must be paid back",
-      "allOf": [
-        {
-          "$ref": "#/definitions/AssetInfo"
-        }
-      ]
-    },
-    "flash_loan_fee": {
-      "description": "The amount of fee paid to vault holders",
-      "allOf": [
-        {
-          "$ref": "#/definitions/Uint128"
-        }
-      ]
-    },
-    "payback_amount": {
-      "description": "The total amount that must be returned. Equivalent to `amount` + `protocol_fee` + `flash_loan_fee`.",
-      "allOf": [
-        {
-          "$ref": "#/definitions/Uint128"
-        }
-      ]
-    },
-    "protocol_fee": {
-      "description": "The amount of fee paid to the protocol",
-      "allOf": [
-        {
-          "$ref": "#/definitions/Uint128"
-        }
-      ]
-    }
-  },
-  "additionalProperties": false,
-  "definitions": {
-    "AssetInfo": {
-      "description": "AssetInfo contract_addr is usually passed from the cw20 hook so we can trust the contract_addr is properly validated.",
-      "oneOf": [
-        {
-          "type": "object",
-          "required": [
-            "token"
-          ],
-          "properties": {
-            "token": {
-              "type": "object",
-              "required": [
-                "contract_addr"
-              ],
-              "properties": {
-                "contract_addr": {
-                  "type": "string"
-                }
-              },
-              "additionalProperties": false
-            }
-          },
-          "additionalProperties": false
-        },
-        {
-          "type": "object",
-          "required": [
-            "native_token"
-          ],
-          "properties": {
-            "native_token": {
-              "type": "object",
-              "required": [
-                "denom"
-              ],
-              "properties": {
-                "denom": {
-                  "type": "string"
-                }
-              },
-              "additionalProperties": false
-            }
-          },
-          "additionalProperties": false
-        }
-      ]
-    },
-    "Uint128": {
-      "description": "A thin wrapper around u128 that is using strings for JSON encoding/decoding, such that the full u128 range can be used for clients that convert JSON numbers to floats, like JavaScript and jq.\n\n# Examples\n\nUse `from` to create instances of this and `u128` to get the value out:\n\n``` # use cosmwasm_std::Uint128; let a = Uint128::from(123u128); assert_eq!(a.u128(), 123);\n\nlet b = Uint128::from(42u64); assert_eq!(b.u128(), 42);\n\nlet c = Uint128::from(70u32); assert_eq!(c.u128(), 70); ```",
-      "type": "string"
-    }
-  }
-}
