--- contracts/liquidity_hub/vault-manager/schema/raw/response_to_vault.json
@@ -1,135 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "title": "VaultsResponse",
-  "description": "Response for the vaults query",
-  "type": "object",
-  "required": [
-    "vaults"
-  ],
-  "properties": {
-    "vaults": {
-      "type": "array",
-      "items": {
-        "$ref": "#/definitions/Vault"
-      }
-    }
-  },
-  "additionalProperties": false,
-  "definitions": {
-    "AssetInfo": {
-      "description": "AssetInfo contract_addr is usually passed from the cw20 hook so we can trust the contract_addr is properly validated.",
-      "oneOf": [
-        {
-          "type": "object",
-          "required": [
-            "token"
-          ],
-          "properties": {
-            "token": {
-              "type": "object",
-              "required": [
-                "contract_addr"
-              ],
-              "properties": {
-                "contract_addr": {
-                  "type": "string"
-                }
-              },
-              "additionalProperties": false
-            }
-          },
-          "additionalProperties": false
-        },
-        {
-          "type": "object",
-          "required": [
-            "native_token"
-          ],
-          "properties": {
-            "native_token": {
-              "type": "object",
-              "required": [
-                "denom"
-              ],
-              "properties": {
-                "denom": {
-                  "type": "string"
-                }
-              },
-              "additionalProperties": false
-            }
-          },
-          "additionalProperties": false
-        }
-      ]
-    },
-    "Decimal": {
-      "description": "A fixed-point decimal value with 18 fractional digits, i.e. Decimal(1_000_000_000_000_000_000) == 1.0\n\nThe greatest possible value that can be represented is 340282366920938463463.374607431768211455 (which is (2^128 - 1) / 10^18)",
-      "type": "string"
-    },
-    "Fee": {
-      "type": "object",
-      "required": [
-        "share"
-      ],
-      "properties": {
-        "share": {
-          "$ref": "#/definitions/Decimal"
-        }
-      },
-      "additionalProperties": false
-    },
-    "Vault": {
-      "description": "Vault representation",
-      "type": "object",
-      "required": [
-        "asset_info",
-        "fees",
-        "lp_asset"
-      ],
-      "properties": {
-        "asset_info": {
-          "description": "The asset info the vault manages",
-          "allOf": [
-            {
-              "$ref": "#/definitions/AssetInfo"
-            }
-          ]
-        },
-        "fees": {
-          "description": "The fees associated with the vault",
-          "allOf": [
-            {
-              "$ref": "#/definitions/VaultFee"
-            }
-          ]
-        },
-        "lp_asset": {
-          "description": "The LP asset",
-          "allOf": [
-            {
-              "$ref": "#/definitions/AssetInfo"
-            }
-          ]
-        }
-      },
-      "additionalProperties": false
-    },
-    "VaultFee": {
-      "type": "object",
-      "required": [
-        "flash_loan_fee",
-        "protocol_fee"
-      ],
-      "properties": {
-        "flash_loan_fee": {
-          "$ref": "#/definitions/Fee"
-        },
-        "protocol_fee": {
-          "$ref": "#/definitions/Fee"
-        }
-      },
-      "additionalProperties": false
-    }
-  }
-}
